Rewrite with insta snapshot testing
