--- espr/tests/cases/.gitignore
@@ -1,1 +1,0 @@
-*.generated.rs
