--- espr/tests/cases/simple.exp
@@ -1,11 +1,0 @@
-SCHEMA test_schema;
-  ENTITY a;
-    x: REAL;
-    y: REAL;
-  END_ENTITY;
-
-  ENTITY b;
-    z: REAL;
-    a: a;
-  END_ENTITY;
-END_SCHEMA;
